# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: M2 - Atención Ciudadana API
  description: API REST del módulo M2 de Atención Ciudadana, Reclamos y Solicitudes.
  version: 1.0.0
servers:
  - url: /
    description: Mismo origen que sirve Swagger UI
...
paths:
...
  /api/tickets:
    get:
      tags: [ Tickets ]
      summary: Listar tickets con filtros
      description: |
        Bandeja de tickets para agentes/administradores, con filtros combinables
        por categoría, prioridad, barrio, área responsable y estado (Story 3.1).
      operationId: listTickets
      security:
        - bearerAuth: [ ]
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Priority'
        - name: neighborhoodId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: responsibleAreaId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          description: Propiedades ordenables de Ticket (whitelist propia del backend); formato `campo,direccion`.
          schema:
            type: array
            items:
              type: string
          example: [ "createdAt,desc" ]
      responses:
        '200':
          description: Página de tickets que cumplen los filtros.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTicketResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Tickets]
      summary: Crear un ticket
      description: |
        Crea un ticket autenticado. La misma operación acepta JSON puro o multipart.

        En JSON no se adjuntan archivos; si el riesgo calculado internamente es HIGH o CRITICAL,
        responde 422 EVIDENCE_REQUIRED. En multipart, `data` es un part JSON obligatorio y
        `evidence` admite hasta 5 archivos opcionales, de 10 MB cada uno y 50 MB en total.
        MIME admitidos: image/jpeg, image/png, image/webp, image/heic, image/heif y application/pdf.
      operationId: createTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
          multipart/form-data:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  $ref: '#/components/schemas/CreateTicketRequest'
                evidence:
                  type: array
                  maxItems: 5
                  description: Archivos opcionales; máximo 10 MB por archivo y 50 MB en total.
                  items:
                    type: string
                    format: binary
            encoding:
              data:
                contentType: application/json
              evidence:
                contentType: image/jpeg, image/png, image/webp, image/heic, image/heif, application/pdf
      responses:
        '201':
          description: Ticket creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTicketResponse'
        '400':
          description: Request inválido, part `data` ausente o adjunto inválido. INVALID_ATTACHMENT sólo aplica a multipart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                missingData:
                  value:
                    code: INVALID_REQUEST
                    message: Falta el part obligatorio 'data'
                invalidAttachment:
                  value:
                    code: INVALID_ATTACHMENT
                    message: Cada evidencia debe contener un archivo no vacío.
        '401':
          $ref: '#/components/responses/InvalidToken'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: Límite de adjuntos excedido; sólo aplica a multipart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: PAYLOAD_TOO_LARGE
                message: La evidencia supera el tamaño máximo permitido.
        '415':
          description: Content-Type del request no soportado o MIME de evidencia no admitido. El MIME de evidencia sólo aplica a multipart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: UNSUPPORTED_MEDIA_TYPE
                message: Tipo de contenido no soportado
        '422':
          description: El riesgo calculado requiere al menos una evidencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: EVIDENCE_REQUIRED
                message: El nivel de riesgo detectado requiere adjuntar al menos una evidencia.
        '503':
          description: El almacenamiento de adjuntos no está disponible; sólo aplica cuando multipart contiene evidencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: ATTACHMENT_STORAGE_UNAVAILABLE
                message: No fue posible almacenar la evidencia. Intentá nuevamente más tarde.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token Bearer obtenido del proveedor de identidad vigente. En DEV puede obtenerse mediante /api/auth/login.
...
  responses:
    InvalidRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_REQUEST
            message: El request es inválido
    InvalidToken:
      description: Token Bearer ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_TOKEN
            message: La sesión no es válida
    Forbidden:
      description: La identidad autenticada no tiene permiso para realizar la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: FORBIDDEN
            message: No está autorizado para realizar esta operación
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: NOT_FOUND
            message: Recurso no encontrado
...
  schemas:
    TicketPublicId:
      type: string
      description: Identificador humano estable del ticket; no reemplaza al UUID ni al código secreto de tracking.
      pattern: '^TK-[0-9]{4}-[0-9]{6,}$'
      example: TK-2026-000123
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
...
    CreateTicketRequest:
      type: object
      required: [requestTypeId, summary, description, formData]
      properties:
        requestTypeId:
          type: integer
          format: int64
        summary:
          type: string
          maxLength: 200
        description:
          type: string
        formData:
          type: object
          description: Objeto dinámico cuyas keys corresponden al `code` de los campos del formulario vigente.
          additionalProperties: true
        location:
          allOf:
            - $ref: '#/components/schemas/LocationData'
          nullable: true
      additionalProperties: false
    LocationData:
      type: object
      properties:
        addressLine:
          type: string
          maxLength: 300
          nullable: true
        street:
          type: string
          maxLength: 150
          nullable: true
        streetNumber:
          type: string
          maxLength: 30
          nullable: true
        neighborhoodId:
          type: string
          format: uuid
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        reference:
          type: string
          maxLength: 500
          nullable: true
    TicketResponse:
      type: object
      description: |
        Vista compartida de un ticket para listados y transiciones. Los GET de
        detalle usan TicketDetailResponse. No confundir con TrackingTicketResponse,
        que es la vista pública reducida. Los deadlines SLA son proyecciones
        derivadas de TicketSla y no atributos persistidos del Ticket.
      required: [id, publicId, requestTypeCode, requestTypeName, categoryName, subcategoryName,
                 ticketType, summary, currentStatus, currentPriority, responsibleAreaId,
                 assignedAgentId, anonymous, estimatedAffectedCount, escalated, neighborhoodId,
                 escalationReasonCode, escalatedAt, firstResponseDueAt, firstResponseNearDue,
                 firstResponseBreached, resolutionDueAt, slaNearDue, slaBreached, resolutionNearDueAt,
                 neighborhoodName, classificationFinalizedAt,
                 statusChangedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        publicId:
          $ref: '#/components/schemas/TicketPublicId'
        requestTypeCode:
          type: string
        requestTypeName:
          type: string
        categoryName:
          type: string
        subcategoryName:
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        summary:
          type: string
        currentStatus:
          $ref: '#/components/schemas/TicketStatus'
        currentPriority:
          $ref: '#/components/schemas/Priority'
        responsibleAreaId:
          type: string
        assignedAgentId:
          type: string
          description: Id interno (ModuleUser.id) del agente asignado, si existe.
          nullable: true
        anonymous:
          type: boolean
        estimatedAffectedCount:
          type: integer
          format: int32
        escalated:
          type: boolean
        escalationReasonCode:
          allOf:
            - $ref: '#/components/schemas/EscalationReasonCode'
          nullable: true
        escalatedAt:
          type: string
          format: date-time
          nullable: true
        firstResponseDueAt:
          type: string
          format: date-time
          nullable: true
          description: Deadline del último ciclo FIRST_RESPONSE; se conserva como referencia histórica al finalizar.
        firstResponseNearDue:
          type: boolean
          description: El último ciclo FIRST_RESPONSE se encuentra actualmente en estado NEAR_DUE.
        firstResponseBreached:
          type: boolean
          description: El último ciclo FIRST_RESPONSE está en BREACHED, incluso si ya finalizó.
        resolutionDueAt:
          type: string
          format: date-time
          nullable: true
          description: Deadline del último ciclo RESOLUTION, derivado desde TicketSla.
        slaNearDue:
          type: boolean
          description: El ciclo de resolución más reciente se encuentra actualmente en estado NEAR_DUE.
        slaBreached:
          type: boolean
          description: El ciclo de resolución incumplió su SLA.
        resolutionNearDueAt:
          type: string
          format: date-time
          nullable: true
          description: Instante calculado con calendario para el umbral preventivo del ciclo de resolución actual o más reciente.
        neighborhoodId:
          type: string
          format: uuid
          nullable: true
        neighborhoodName:
          type: string
          nullable: true
        classificationFinalizedAt:
          type: string
          format: date-time
          nullable: true
        statusChangedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    PagedTicketResponse:
      type: object
      description: |
        Forma simplificada de la página que devuelve Spring Data (`Page<TicketResponse>`)
        para el listado de tickets; se documentan los campos estables para el
        cliente, sin los metadatos internos de paginación de Spring (`pageable`, `sort`).
      required: [content, totalElements, totalPages, size, number, first, last, numberOfElements, empty]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
          description: Página actual (0-indexed).
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...
    CreateTicketResponse:
      type: object
      required: [ticketId, publicId, trackingCode, status]
      properties:
        ticketId:
          type: string
          format: uuid
        publicId:
          $ref: '#/components/schemas/TicketPublicId'
        trackingCode:
          type: string
          description: Código que permite consultar el ticket mediante el endpoint público de tracking.
        status:
          $ref: '#/components/schemas/TicketStatus'
...
    TicketType:
      type: string
      enum: [COMPLAINT, REQUEST, INQUIRY, SUGGESTION]
    TicketStatus:
      type: string
      enum: [REGISTERED, IN_REVIEW, ROUTED, IN_PROGRESS, PENDING_INFORMATION, DUPLICATE, RESOLVED, CLOSED, CANCELLED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    EscalationReasonCode:
      type: string
      enum: [CRITICAL_PRIORITY, SLA_BREACHED, MANUAL]
...